+++ tests/Unit/Consumer/Fixture/FakeClient.php
@@ -435,7 +435,8 @@ final class FakeClient extends Client
         string $memberId,
         int $generationId,
         array $topicPartitionOffsets,
-        int $retentionTimeMs
+        int $retentionTimeMs,
+        ?string $groupInstanceId = null
     ): void {
         $failure = array_shift($this->commitFailures);
         if ($failure !== null) {
@@ -448,6 +449,7 @@ final class FakeClient extends Client
             'generationId'  => $generationId,
             'offsets'       => $topicPartitionOffsets,
             'retentionTime' => $retentionTimeMs,
+            'instanceId'    => $groupInstanceId,
         ];
 
         foreach ($topicPartitionOffsets as $topic => $partitionOffsets) {
@@ -468,7 +470,8 @@ final class FakeClient extends Client
         string $memberId,
         string $protocolType,
         array $groupProtocols,
-        ?int $rebalanceTimeoutMs = null
+        ?int $rebalanceTimeoutMs = null,
+        ?string $groupInstanceId = null
     ): JoinGroupResponse {
         $this->joins[] = [
             'groupId'          => $groupId,
@@ -476,6 +479,7 @@ final class FakeClient extends Client
             'protocolType'     => $protocolType,
             'protocols'        => $groupProtocols,
             'rebalanceTimeout' => $rebalanceTimeoutMs,
+            'instanceId'       => $groupInstanceId,
         ];
 
         $failure = array_shift($this->joinFailures);
@@ -519,13 +523,15 @@ final class FakeClient extends Client
         string $groupId,
         string $memberId,
         int $generationId,
-        array $groupAssignments = []
+        array $groupAssignments = [],
+        ?string $groupInstanceId = null
     ): SyncGroupResponse {
         $this->syncs[] = [
             'groupId'      => $groupId,
             'memberId'     => $memberId,
             'generationId' => $generationId,
             'assignments'  => $groupAssignments,
+            'instanceId'   => $groupInstanceId,
         ];
 
         $failure = array_shift($this->syncFailures);
@@ -547,12 +553,18 @@ final class FakeClient extends Client
     /**
      * @inheritdoc
      */
-    public function heartbeat(Node $coordinatorNode, string $groupId, string $memberId, int $generationId): void
-    {
+    public function heartbeat(
+        Node $coordinatorNode,
+        string $groupId,
+        string $memberId,
+        int $generationId,
+        ?string $groupInstanceId = null
+    ): void {
         $this->heartbeats[] = [
             'groupId'      => $groupId,
             'memberId'     => $memberId,
             'generationId' => $generationId,
+            'instanceId'   => $groupInstanceId,
         ];
 
         $failure = array_shift($this->heartbeatFailures);